+++ cross/applications/pages/steps/spline_transformation/spline_transformation.py
@@ -1,0 +1,5 @@
+class SplineTransformationsBase:
+    TRANSFORMATIONS = {
+        "Do nothing": "none",
+        "Spline": "spline",
+    }